Module rust_final_project::transform

This module provides functions to transform movie data from a list of Movie structs to a list of TransformedMovie structs. The transformation process includes converting the data to a DataFrame, cleaning and validating the data, and then converting it back to the desired format.

Functions

clean_and_validate_data 🔒

Cleans and validates the data in a DataFrame.

dataframe_to_transformed_movies 🔒

Converts a cleaned DataFrame into a vector of TransformedMovie structs.

movies_to_dataframe 🔒

Converts a vector of Movie structs into a Polars DataFrame.

transform_movies

Transforms a list of Movie structs into a list of TransformedMovie structs.